EDU-1218 Include media tags in redaction page tags tab
