+++ src/components/pages/redaction.js
@@ -4,15 +4,15 @@ import React, { useState } from 'react';
 import routes from '../../utils/routes.js';
 import { useTranslation } from 'react-i18next';
 import { useRequest } from '../request-context.js';
+import RedactionTagsTab from '../redaction/redaction-tags-tab.js';
 import RedactionDocumentsTab from '../redaction/redaction-documents-tab.js';
-import RedactionDocumentTagsTab from '../redaction/redaction-document-tags-tab.js';
 import RedactionMediaLibraryTab from '../redaction/redaction-media-library-tab.js';
 import { documentExtendedMetadataShape, mediaLibraryItemShape } from '../../ui/default-prop-types.js';
 
 const TABS = {
   documents: 'documents',
-  documentTags: 'document-tags',
-  mediaLibrary: 'media-library'
+  mediaLibrary: 'media-library',
+  tags: 'tags'
 };
 
 const determineTab = query => Object.values(TABS).find(val => val === query) || Object.keys(TABS)[0];
@@ -44,20 +44,20 @@ function Redaction({ initialState, PageTemplate }) {
       )
     },
     {
-      key: TABS.documentTags,
-      label: t('documentTagsTabTitle'),
-      children: (
-        <div className="Tabs-tabPane">
-          <RedactionDocumentTagsTab documents={documents} />
-        </div>
-      )
-    },
-    {
       key: TABS.mediaLibrary,
       label: t('mediaLibraryTabTitle'),
       children: (
         <div className="Tabs-tabPane">
           <RedactionMediaLibraryTab mediaLibraryItems={mediaLibraryItems} onMediaLibraryItemsChange={setMediaLibraryItems} />
+        </div>
+      )
+    },
+    {
+      key: TABS.tags,
+      label: t('tagsTabTitle'),
+      children: (
+        <div className="Tabs-tabPane">
+          <RedactionTagsTab documents={documents} mediaLibraryItems={mediaLibraryItems} />
         </div>
       )
     }
