--- src/components/redaction/redaction-document-tags-tab.less
@@ -1,52 +1,0 @@
-.RedactionDocumentTagsTab {}
-
-.RedactionDocumentTagsTab-controls {
-  margin-bottom: 40px;
-  display: grid;
-  gap: 15px;
-  grid-template-columns: 2fr 1fr;
-
-  @media @edu-media-phone {
-    grid-template-columns: 100%;
-    grid-template-rows: 1fr 1fr;
-  }
-}
-
-.RedactionDocumentTagsTab-filter {
-  width: 100%;
-}
-
-.RedactionDocumentTagsTab-expandedRow {
-  font-size: @edu-font-size-sm;
-}
-
-.RedactionDocumentTagsTab-expandedRowHeader {
-  font-weight: bold;
-  margin-bottom: 5px;
-
-  &:not(:first-child) {
-    margin-top: 20px;
-  }
-}
-
-.RedactionDocumentTagsTab-documentList {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
-
-.RedactionDocumentTagsTab-companionTags {
-  line-height: 200%;
-}
-
-.RedactionDocumentTagsTab-companionTag {
-  white-space: nowrap;
-
-  &:not(:last-child) {
-    margin-right: 15px;
-  }
-}
-
-.RedactionDocumentTagsTab-companionTagFrequency {
-  font-size: @edu-font-size-xs;
-}
