+++ src/styles/main.less
@@ -90,6 +90,7 @@
 @import '../components/pages/browser-not-supported.less';
 @import '../components/media-player/youtube-player.less';
 @import '../components/irreversible-actions-section.less';
+@import '../components/redaction/redaction-tags-tab.less';
 @import '../components/pages/connect-external-account.less';
 @import '../components/homepage/homepage-trust-footer.less';
 @import '../components/media-player/track-mixer-editor.less';
@@ -108,7 +109,6 @@
 @import '../components/resource-selector/resource-selector.less';
 @import '../components/admin/homepage-presentation-settings.less';
 @import '../components/media-player/multitrack-media-player.less';
-@import '../components/redaction/redaction-document-tags-tab.less';
 @import '../components/redaction/redaction-media-library-tab.less';
 @import '../components/resource-selector/shared/resource-url.less';
 @import '../components/admin/user-account-locked-state-editor.less';
